Change vehicleBreadth in subclass of vehicle to int. Add getter/setter method for roadWidth and test. Work on paint road method.

+++ src/Simulation/Bus.java
@@ -11,7 +11,7 @@ public class Bus extends Vehicle {
      * @param vehicleBreadth the vehicle breadth - Half of its length
      */
 
-    public Bus(int speed, int positionOnRoad, int id, float vehicleBreadth) {
+    public Bus(int speed, int positionOnRoad, int id, int vehicleBreadth) {
         super(speed, positionOnRoad, id, vehicleBreadth);
     }
