+++ src/Simulation/Car.java
@@ -13,7 +13,7 @@ public class Car extends Vehicle {
      * @param vehicleBreadth the vehicle breadth - Half of its length
      */
 
-    public Car(int speed, int positionOnRoad, int id, float vehicleBreadth) {
+    public Car(int speed, int positionOnRoad, int id, int vehicleBreadth) {
         super(speed, positionOnRoad, id, vehicleBreadth);
     }
